frontend/requirements.txt
# requirements.txt for frontend (Streamlit UI)
streamlit
pandas
requests
python-dotenv